requirements.txt
Pillow
newspaper3k
urllib3
streamlit
beautifulsoup4
protobuf==3.20.0
jinja2>=3.0
wordcloud 
matplotlib
textblob
nltk
streamlit-star-rating
pymongo
